Test/requirements.txt
numpy
pyserial
websockets==12.0
